change contribution implementation to use float data type instead of double, inorder to comply with the kotlin-ipv8 module

+++ musicdao/src/main/java/nl/tudelft/trustchain/musicdao/core/contribute/Contribution.kt
@@ -4,6 +4,6 @@ import nl.tudelft.trustchain.musicdao.core.repositories.model.Artist
 
 data class Contribution(
     val id: String,
-    val amount: Double,
+    val amount: Float,
     val artists: List<Artist>
 )
